rename 'multi_process_*' -> 'multiprocess' for consistency
