+++ allennlp/data/data_loaders/__init__.py
@@ -1,5 +1,5 @@
 from allennlp.data.data_loaders.data_loader import DataLoader, TensorDict, allennlp_collate
-from allennlp.data.data_loaders.multi_process_data_loader import MultiProcessDataLoader, WorkerError
+from allennlp.data.data_loaders.multiprocess_data_loader import MultiProcessDataLoader, WorkerError
 from allennlp.data.data_loaders.multitask_data_loader import MultiTaskDataLoader
 
 from allennlp.data.data_loaders.pytorch_data_loader import (
